use a template and compile
`build.py` will compile all *.tpl.html files in the _posts/_
directory in addition to _index.tpl.html_ using _template.tpl.html_.
The first line in each template is the title of the page.

The process is an ugly hack and hardcodes some things that shouldn't
be hardcoded, but is a good enough solution for now. (It's better than
the current solution of copy/paste.)

*TODO*: It would be nice to automate _index.html_ to include the list
of blog posts at some point.